Added fine servo adjustment, started calibration

+++ Lab10/servo.h
@@ -10,6 +10,6 @@
 
 void servo_init(void);
 
-void servo_move(uint16_t degrees);
+void servo_move(int32_t degrees);
 
 #endif /* SERVO_H_ */
